Messing with new walls, added regenerate button, added point light that follows the player. Also moved all game logic into new module
