+++ assets/meshes/mazeWall.obj
@@ -1,0 +1,180 @@
+# Blender v2.92.0 OBJ File: 'mazeWall.blend'
+# www.blender.org
+mtllib mazeWall.mtl
+o Cube_Cube.001
+v -1.000070 -1.000000 1.000070
+v -1.000070 0.942898 1.000070
+v -1.000070 -1.000000 -1.000070
+v -1.000070 0.942898 -1.000070
+v 1.000070 -1.000000 1.000070
+v 1.000070 0.942898 1.000070
+v 1.000070 -1.000000 -1.000070
+v 1.000070 0.942898 -1.000070
+v -1.031000 -1.000000 0.885998
+v -1.031000 0.859357 0.885998
+v -1.031000 0.859357 -0.885998
+v -1.031000 -1.000000 -0.885998
+v -0.885998 -1.000000 -1.031000
+v -0.885998 0.859357 -1.031000
+v 0.885998 0.859357 -1.031000
+v 0.885998 -1.000000 -1.031000
+v 1.031000 -1.000000 -0.885998
+v 1.031000 0.859357 -0.885998
+v 1.031000 0.859357 0.885998
+v 1.031000 -1.000000 0.885998
+v 0.885998 -1.000000 1.031000
+v 0.885998 0.859357 1.031000
+v -0.885998 0.859357 1.031000
+v -0.885998 -1.000000 1.031000
+v 0.761368 0.990178 -0.761368
+v -0.761368 0.990178 -0.761368
+v -0.761368 0.990178 0.761368
+v 0.761368 0.990178 0.761368
+vt 0.929679 0.929679
+vt 0.070321 -0.000000
+vt 0.929679 0.000000
+vt 0.929679 0.929679
+vt 0.070321 0.000000
+vt 0.929679 0.000000
+vt 0.929679 0.929679
+vt 0.070321 0.000000
+vt 0.929679 0.000000
+vt 0.929679 0.929679
+vt 0.070321 0.000000
+vt 0.929679 0.000000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 0.880657 0.119343
+vt 0.119343 0.880657
+vt 0.119343 0.119343
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 1.000000 1.000000
+vt 0.070321 0.929679
+vt 0.375000 0.000000
+vt 0.392580 0.232420
+vt 0.375000 0.250000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 1.000000 1.000000
+vt 0.000000 0.000000
+vt 0.070321 0.929679
+vt 0.392580 0.482420
+vt 0.375000 0.500000
+vt 0.000000 1.000000
+vt 1.000000 1.000000
+vt 0.000000 0.000000
+vt 0.070321 0.929679
+vt 0.392580 0.732420
+vt 0.375000 0.750000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 1.000000 1.000000
+vt 0.000000 0.000000
+vt 0.070321 0.929679
+vt 0.392580 0.982420
+vt 0.375000 1.000000
+vt 1.000000 0.000000
+vt 0.000000 0.000000
+vt 0.880657 0.880657
+vt 1.000000 1.000000
+vt 0.392580 0.017580
+vt 0.392580 0.267580
+vt 0.392580 0.517580
+vt 0.392580 0.767580
+vn -0.9855 0.1412 0.0939
+vn -0.9912 0.0000 -0.1320
+vn -0.9912 0.0000 0.1320
+vn -0.0939 0.1412 -0.9855
+vn 0.1320 0.0000 -0.9912
+vn -0.1320 0.0000 -0.9912
+vn 0.9855 0.1412 -0.0939
+vn 0.9912 0.0000 0.1320
+vn 0.9912 0.0000 -0.1320
+vn 0.0939 0.1412 0.9855
+vn -0.1320 0.0000 0.9912
+vn 0.1320 0.0000 0.9912
+vn 0.0000 -1.0000 -0.0000
+vn -0.0734 0.9946 -0.0734
+vn 0.0734 0.9946 0.0734
+vn 0.0734 0.9946 -0.0734
+vn -0.9652 0.0000 0.2617
+vn -0.9766 0.1456 -0.1582
+vn -0.9766 0.1456 0.1582
+vn -0.9652 0.0000 -0.2617
+vn -0.9855 0.1412 -0.0939
+vn -0.2617 0.0000 -0.9652
+vn 0.1582 0.1456 -0.9766
+vn -0.1582 0.1456 -0.9766
+vn 0.2617 0.0000 -0.9652
+vn 0.0939 0.1412 -0.9855
+vn 0.9652 0.0000 -0.2617
+vn 0.9766 0.1456 0.1582
+vn 0.9766 0.1456 -0.1582
+vn 0.9652 0.0000 0.2617
+vn 0.9855 0.1412 0.0939
+vn 0.2617 0.0000 0.9652
+vn -0.1582 0.1456 0.9766
+vn 0.1582 0.1456 0.9766
+vn -0.2617 0.0000 0.9652
+vn -0.0939 0.1412 0.9855
+vn -0.0981 0.9903 -0.0981
+vn 0.0981 0.9903 -0.0981
+vn -0.0981 0.9903 0.0981
+vn 0.0981 0.9903 0.0981
+vn -0.0734 0.9946 0.0734
+usemtl None
+s 1
+f 10/1/1 12/2/2 9/3/3
+f 14/4/4 16/5/5 13/6/6
+f 18/7/7 20/8/8 17/9/9
+f 22/10/10 24/11/11 21/12/12
+f 7/13/13 1/14/13 3/15/13
+f 26/16/14 28/17/15 25/18/16
+f 1/19/17 10/1/1 9/3/3
+f 4/20/18 10/1/1 2/21/19
+f 3/15/20 11/22/21 4/20/18
+f 1/23/13 12/24/13 3/25/13
+f 3/26/22 14/4/4 13/6/6
+f 8/27/23 14/4/4 4/28/24
+f 7/29/25 15/30/26 8/27/23
+f 3/25/13 16/31/13 7/32/13
+f 7/13/27 18/7/7 17/9/9
+f 6/33/28 18/7/7 8/34/29
+f 5/35/30 19/36/31 6/33/28
+f 7/32/13 20/37/13 5/38/13
+f 5/39/32 22/10/10 21/12/12
+f 2/40/33 22/10/10 6/41/34
+f 1/42/35 23/43/36 2/40/33
+f 5/38/13 24/44/13 1/45/13
+f 4/46/37 25/18/16 8/47/38
+f 2/21/39 26/16/14 4/46/37
+f 6/33/40 27/48/41 2/21/39
+f 8/47/38 28/17/15 6/33/40
+f 10/1/1 11/22/21 12/2/2
+f 14/4/4 15/30/26 16/5/5
+f 18/7/7 19/36/31 20/8/8
+f 22/10/10 23/43/36 24/11/11
+f 7/13/13 5/49/13 1/14/13
+f 26/16/14 27/48/41 28/17/15
+f 1/19/17 2/21/19 10/1/1
+f 4/20/18 11/22/21 10/1/1
+f 3/15/20 12/2/2 11/22/21
+f 1/23/13 9/50/13 12/24/13
+f 3/26/22 4/28/24 14/4/4
+f 8/27/23 15/30/26 14/4/4
+f 7/29/25 16/5/5 15/30/26
+f 3/25/13 13/51/13 16/31/13
+f 7/13/27 8/34/29 18/7/7
+f 6/33/28 19/36/31 18/7/7
+f 5/35/30 20/8/8 19/36/31
+f 7/32/13 17/52/13 20/37/13
+f 5/39/32 6/41/34 22/10/10
+f 2/40/33 23/43/36 22/10/10
+f 1/42/35 24/11/11 23/43/36
+f 5/38/13 21/53/13 24/44/13
+f 4/46/37 26/16/14 25/18/16
+f 2/21/39 27/48/41 26/16/14
+f 6/33/40 28/17/15 27/48/41
+f 8/47/38 25/18/16 28/17/15
