+++ index.html
@@ -45,6 +45,10 @@
         <p id="meshes">- meshes</p>
     </div>
 
+    <button id="regen-button">
+        regenerate
+    </button>
+
 </body>
 
 </html>
