+++ js/app.mjs
@@ -1,5 +1,8 @@
-import startHelloWebGL from "./renderer/renderer.mjs";
+import startGame from "./game.mjs";
+import initWebGL from "./renderer/renderer.mjs";
 
 window.addEventListener("load", (_) => {
-    startHelloWebGL();
+    initWebGL();
+
+    startGame();
 });
